feat: move product page component to it's own folder
