+++ src/App.tsx
@@ -3,7 +3,7 @@ import { useState } from 'react';
 import { User } from './types';
 import Home from './pages/Home/Home';
 import ProductDetails from './pages/ProductDetails';
-import Products from './pages/Products';
+import Products from './pages/Products/Products';
 import Header from './components/Header/Header';
 import Footer from './components/Footer/Footer';
 import Cart from './pages/Cart';
