+++ src/pages/Products/Products.tsx
@@ -1,8 +1,8 @@
 import { useState, useEffect } from 'react';
 import { Col, Container, Row } from 'react-bootstrap';
-import ProductCard from '../components/Cards/ProductCard';
-import Pagination from '../components/Pagination/Pagination';
-import { Product } from '../types';
+import ProductCard from '../../components/Cards/ProductCard';
+import Pagination from '../../components/Pagination/Pagination';
+import { Product } from '../../types';
 
 function Products() {
   const [products, setProducts] = useState<Product[]>([]);
